requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.2.0=eigen
absl-py=0.9.0=py36h9f0ad1d_1
alabaster=0.7.12=py_0
argh=0.26.2=py36_1001
arrow=0.13.1=py36_0
astor=0.7.1=py_0
astroid=2.3.3=py36_1
atomicwrites=1.3.0=py_0
attrs=19.3.0=py_0
autopep8=1.5=py_0
babel=2.8.0=py_0
backcall=0.1.0=py_0
bcrypt=3.1.7=py36hfa6e2cd_0
binaryornot=0.4.4=pyhd3eb1b0_1
blas=1.0=mkl
bleach=3.1.4=pyh9f0ad1d_0
ca-certificates=2020.10.14=0
certifi=2020.6.20=py36_0
cffi=1.14.0=py36ha419a9e_0
chardet=3.0.4=py36h9f0ad1d_1006
click=8.0.1=pyhd3eb1b0_0
cloudpickle=1.3.0=py_0
colorama=0.4.3=py_0
configparser=5.0.2=pypi_0
cookiecutter=1.7.3=pyh6c4a22f_0
coverage=5.3=py36he774522_0
cryptography=2.8=py36hb32ad35_1
cuda90=1.0=0
cuda91=1.0=0
cudatoolkit=10.1.243=h74a9793_0
cycler=0.10.0=py_2
cytoolz=0.10.1=py36hfa6e2cd_0
dask-core=2.13.0=py_0
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
diff-match-patch=20181111=py_0
docker-pycreds=0.4.0=pypi_0
docutils=0.16=py36h9f0ad1d_1
entrypoints=0.3=py36h9f0ad1d_1001
flake8=3.7.9=py36_0
freetype=2.10.1=ha9979f8_0
future=0.18.2=py36h9f0ad1d_1
gast=0.2.2=py_0
geos=3.8.1=he025d50_0
gitdb=4.0.7=pypi_0
gitpython=3.1.17=pypi_0
google-pasta=0.2.0=pyh8c360ce_0
graphviz=2.38.0=h6538335_1011
grpcio=1.23.0=py36hc6b9980_1
h5py=2.10.0=nompi_py36h422b98e_102
hdf5=1.10.5=nompi_ha405e13_1104
helper=2.5.0=pyh44b312d_0
icc_rt=2019.0.0=h0cc432a_1
icu=64.2=he025d50_1
idna=2.9=py_1
imageio=2.8.0=py_0
imagesize=1.2.0=py_0
imgaug=0.4.0=py_1
importlib-metadata=1.5.2=py36h9f0ad1d_0
importlib_metadata=1.5.2=0
iniconfig=1.1.1=py_0
intel-openmp=2020.0=166
intervaltree=3.0.2=py_0
ipykernel=5.2.0=py36h5ca1d4c_1
ipython=7.13.0=py36h9f0ad1d_2
ipython_genutils=0.2.0=py_1
isort=4.3.21=py36h9f0ad1d_1
jedi=0.14.1=py36_0
jinja2=2.11.1=py_0
jinja2-time=0.2.0=pyhd3eb1b0_2
joblib=0.14.1=py_0
jpeg=9c=hfa6e2cd_1001
jsonschema=3.2.0=py36h9f0ad1d_1
jupyter_client=6.1.2=py_0
jupyter_core=4.6.3=py36h9f0ad1d_1
keras=2.2.4=0
keras-applications=1.0.8=py_1
keras-base=2.2.4=py36_0
keras-preprocessing=1.1.0=py_0
keyring=21.1.1=py36h9f0ad1d_2
kiwisolver=1.1.0=py36h246c5b5_1
lazy-object-proxy=1.4.3=py36h68a101e_2
libblas=3.8.0=15_mkl
libcblas=3.8.0=15_mkl
libclang=9.0.1=default_hf44288c_0
liblapack=3.8.0=15_mkl
liblapacke=3.8.0=15_mkl
libopencv=4.2.0=py36_2
libpng=1.6.37=hfe6a214_1
libprotobuf=3.11.4=h1a1b453_0
libsodium=1.0.17=h2fa13f4_0
libspatialindex=1.9.3=he025d50_3
libtiff=4.1.0=h885aae3_6
libwebp=1.0.2=hfa6e2cd_5
lz4-c=1.8.3=he025d50_1001
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.2.1=py_0
markupsafe=1.1.1=py36h68a101e_1
matplotlib=3.2.1=0
matplotlib-base=3.2.1=py36hf0cddfc_0
mccabe=0.6.1=py_1
mistune=0.8.4=py36hfa6e2cd_1000
mkl=2020.0=166
mkl-service=2.3.0=py36hb782905_0
more-itertools=8.5.0=py_0
msys2-conda-epoch=20160418=1
nbconvert=5.6.1=py36_0
nbformat=5.0.4=py_0
networkx=2.4=py_1
ninja=1.10.0=h1ad3211_0
numpy=1.18.1=py36h48dd78f_1
numpydoc=0.9.2=py_0
olefile=0.46=py_0
opencv=4.2.0=py36_2
openssl=1.1.1h=he774522_0
opt_einsum=3.2.0=py_0
packaging=20.1=py_0
pandas=1.1.5=pypi_0
pandoc=2.9.2=0
pandocfilters=1.4.2=py_1
paramiko=2.7.1=py36_0
parso=0.5.2=py_0
pathtools=0.1.2=py_1
pexpect=4.8.0=py36h9f0ad1d_1
pickleshare=0.7.5=py36h9f0ad1d_1001
pillow=7.0.0=py36h8ffffea_1
pip=20.0.2=py_2
pluggy=0.12.0=py_0
poyo=0.5.0=pyhd3eb1b0_0
promise=2.3=pypi_0
prompt-toolkit=3.0.5=py_0
protobuf=3.17.2=pypi_0
psutil=5.7.0=py36h68a101e_1
py=1.9.0=py_0
py-opencv=4.2.0=py36h5ca1d4c_2
pycodestyle=2.5.0=py_0
pycparser=2.20=py_0
pydocstyle=5.0.2=py_0
pydot=1.4.1=py36_0
pyflakes=2.1.1=py_0
pygments=2.6.1=py_0
pylint=2.4.4=py36_0
pynacl=1.3.0=py36h2fa13f4_1001
pyopenssl=19.1.0=py_1
pyparsing=2.4.6=py_0
pyqt=5.12.3=py36h6538335_1
pyqt5-sip=4.19.18=pypi_0
pyqtwebengine=5.12.1=pypi_0
pyreadline=2.1=py36_1001
pyrsistent=0.16.0=py36h68a101e_0
pysocks=1.7.1=py36h9f0ad1d_1
pytest=6.1.1=py36_0
python=3.6.10=h9f7ef89_0
python-dateutil=2.8.1=py_0
python-jsonrpc-server=0.3.4=py_0
python-language-server=0.31.9=py36h9f0ad1d_0
python-slugify=5.0.2=pyhd3eb1b0_0
python_abi=3.6=1_cp36m
pytorch=1.4.0=py3.6_cuda101_cudnn7_0
pytz=2019.3=py_0
pywavelets=1.1.1=py36hc8d92b1_0
pywin32=227=py36hfa6e2cd_0
pywin32-ctypes=0.2.0=py36h9f0ad1d_1001
pyyaml=5.3.1=py36h68a101e_0
pyzmq=19.0.0=py36ha89082f_1
qdarkstyle=2.8.1=pyh9f0ad1d_0
qt=5.12.5=h7ef1ec2_0
qtawesome=0.7.0=py_0
qtconsole=4.7.2=pyh9f0ad1d_0
qtpy=1.9.0=py_0
requests=2.23.0=pyh8c360ce_2
rope=0.16.0=py_0
rtree=0.9.4=py36h9a6d676_1
scikit-image=0.16.2=py36h47e9c7a_0
scikit-learn=0.22.1=py36h6288b17_0
scipy=1.4.1=py36h9439919_0
sentry-sdk=1.1.0=pypi_0
setuptools=46.1.3=py36h9f0ad1d_0
shapely=1.7.0=py36hcd51921_2
shortuuid=1.0.1=pypi_0
six=1.14.0=py_1
smmap=4.0.0=pypi_0
snakeviz=2.0.1=py36_0
snowballstemmer=2.0.0=py_0
sortedcontainers=2.1.0=py_0
sphinx=2.4.4=py_0
sphinxcontrib-applehelp=1.0.2=py_0
sphinxcontrib-devhelp=1.0.2=py_0
sphinxcontrib-htmlhelp=1.0.3=py_0
sphinxcontrib-jsmath=1.0.1=py_0
sphinxcontrib-qthelp=1.0.3=py_0
sphinxcontrib-serializinghtml=1.1.4=py_0
spyder=4.0.1=py36_0
spyder-kernels=1.8.1=py36_0
sqlite=3.30.1=hfa6e2cd_0
subprocess32=3.5.4=pypi_0
tensorboard=1.15.0=py36_0
tensorflow=1.15.0=eigen_py36h932cce6_0
tensorflow-base=1.15.0=eigen_py36h07d2309_0
tensorflow-estimator=1.15.1=pyh2649769_0
termcolor=1.1.0=py_2
testpath=0.4.4=py_0
text-unidecode=1.3=py_0
tk=8.6.10=hfa6e2cd_0
toml=0.10.1=py_0
toolz=0.10.0=py_0
torch-tb-profiler=0.1.0=pypi_0
torchvision=0.5.0=py36_cu101
tornado=6.0.4=py36hfa6e2cd_0
tqdm=4.42.1=py_0
traitlets=4.3.3=py36h9f0ad1d_1
typed-ast=1.4.1=py36hfa6e2cd_0
typing-extensions=3.10.0.0=pypi_0
ujson=1.35=py36hbd05846_1002
unidecode=1.2.0=pyhd3eb1b0_0
urllib3=1.25.7=py36h9f0ad1d_1
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_1
wandb=0.10.31=pypi_0
watchdog=0.10.2=py36_0
wcwidth=0.1.9=pyh9f0ad1d_0
webencodings=0.5.1=py_1
werkzeug=0.16.1=py_0
wheel=0.34.2=py_1
win_inet_pton=1.1.0=py36_0
wincertstore=0.2=py36_1003
wrapt=1.12.1=py36h68a101e_1
xz=5.2.4=h2fa13f4_1002
yaml=0.2.2=hfa6e2cd_1
yapf=0.31.0=pyhd8ed1ab_0
zeromq=4.3.2=h6538335_2
zipp=3.1.0=py_0
zlib=1.2.11=h2fa13f4_1006
zstd=1.4.4=hd8a0e53_2
